+++ memote/support/consistency_helpers.py
@@ -99,7 +99,7 @@ def stoichiometry_matrix(metabolites, reactions):
 
 def rank(matrix, atol=1e-13, rtol=0):
     """
-    Estimate the rank, i.e. the dimension of the column space, of a matrix.
+    Estimate the rank, i.e., the dimension of the column space, of a matrix.
 
     The algorithm used by this function is based on the singular value
     decomposition of `stoichiometry_matrix`.